+++ noodles-vcf/src/lazy.rs
@@ -1,0 +1,5 @@
+//! Lazily-evaluated VCF record.
+
+pub mod record;
+
+pub use self::record::Record;